backend/requirements.txt
Flask==2.2.5
pytesseract==0.3.10
Pillow==9.5.0
pdf2image==1.16.3
werkzeug==2.2.3
python-dotenv==1.0.0
opencv-python-headless==4.8.0.76
google-cloud-vision==3.5.0 # optional, nếu dùng Google Vision
gunicorn==20.1.0 # production